Relocate misplaced tests in data-model from json-encoding-context to fabric-native-instances
Three describe groups in `packages/data-model/test/json-encoding-context.test.ts`
were exercising things outside the JsonEncodingContext concern:

* `FabricInstance instanceof checks` (9 tests) — protocol shape of the
  FabricInstance class hierarchy.
* `UnknownValue` / `ProblematicValue` / `ExplicitTagValue` (7 tests)
  — class-shape assertions (typeTag/state/error preservation,
  DECONSTRUCT method, instanceof relationships).
* `nativeFromFabricValueModern` (10 tests) — frozen/unfrozen output,
  sparse holes, FabricMap → FrozenMap, FabricSet → FrozenSet, deep
  Error-internals unwrap (C2).

These now live in `packages/data-model/test/fabric-native-instances.test.ts`,
where the related fabric-native-instances class hierarchy and the
existing `nativeFromFabricValueModern` describe block already live.
The two `nativeFromFabricValueModern` describe blocks are merged into
one cohesive top-level describe; constituent it-blocks unchanged in
shape and order.

Imports in `json-encoding-context.test.ts` cleaned up to drop symbols
no longer used after the relocation.
