+++ packages/data-model/test/json-encoding-context.test.ts
@@ -1,24 +1,12 @@
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
 import { JsonEncodingContext } from "../json-encoding-context.ts";
-import {
-  DECONSTRUCT,
-  FabricInstance,
-  type FabricValue,
-  type ReconstructionContext,
-} from "../interface.ts";
+import type { FabricValue, ReconstructionContext } from "../interface.ts";
 import type { JsonWireValue } from "../json-type-handlers.ts";
 import { UnknownValue } from "../unknown-value.ts";
 import { ProblematicValue } from "../problematic-value.ts";
-import { ExplicitTagValue } from "../explicit-tag-value.ts";
 import { FabricEpochDays, FabricEpochNsec } from "../fabric-epoch.ts";
-import {
-  FabricError,
-  FabricMap,
-  FabricSet,
-} from "../fabric-native-instances.ts";
-import { nativeFromFabricValueModern } from "../fabric-value-modern.ts";
-import { FrozenMap, FrozenSet } from "../frozen-builtins.ts";
+import { FabricError } from "../fabric-native-instances.ts";
 import {
   resetDataModelConfig,
   setDataModelConfig,
@@ -1586,131 +1574,6 @@ describe("JsonEncodingContext", () => {
   });
 
   // --------------------------------------------------------------------------
-  // data-model-protocol: FabricInstance instanceof checks
-  // --------------------------------------------------------------------------
-
-  describe("FabricInstance instanceof checks", () => {
-    it("returns false for null", () => {
-      expect((null as unknown) instanceof FabricInstance).toBe(false);
-    });
-
-    it("returns false for undefined", () => {
-      expect((undefined as unknown) instanceof FabricInstance).toBe(false);
-    });
-
-    it("returns false for primitives", () => {
-      expect((42 as unknown) instanceof FabricInstance).toBe(false);
-      expect(("hello" as unknown) instanceof FabricInstance).toBe(false);
-      expect((true as unknown) instanceof FabricInstance).toBe(false);
-    });
-
-    it("returns false for plain objects", () => {
-      expect(({} as unknown) instanceof FabricInstance).toBe(false);
-      expect(({ a: 1 } as unknown) instanceof FabricInstance).toBe(false);
-    });
-
-    it("returns true for UnknownValue", () => {
-      const us = new UnknownValue("Test@1", null);
-      expect(us instanceof FabricInstance).toBe(true);
-    });
-
-    it("returns true for ProblematicValue", () => {
-      const ps = new ProblematicValue("Test@1", null, "oops");
-      expect(ps instanceof FabricInstance).toBe(true);
-    });
-
-    it("returns true for custom FabricInstance subclass", () => {
-      class CustomFabInst extends FabricInstance {
-        [DECONSTRUCT](): FabricValue {
-          return { value: 42 };
-        }
-        protected shallowUnfrozenClone(): CustomFabInst {
-          return new CustomFabInst();
-        }
-      }
-      const instance = new CustomFabInst();
-      expect(instance instanceof FabricInstance).toBe(true);
-    });
-
-    it("returns true for FabricError", () => {
-      const se = new FabricError(new Error("test"));
-      expect(se instanceof FabricInstance).toBe(true);
-    });
-  });
-
-  // --------------------------------------------------------------------------
-  // UnknownValue
-  // --------------------------------------------------------------------------
-
-  describe("UnknownValue", () => {
-    it("preserves typeTag and state", () => {
-      const us = new UnknownValue("FancyType@3", { data: [1, 2, 3] });
-      expect(us.typeTag).toBe("FancyType@3");
-      expect(us.state).toEqual({ data: [1, 2, 3] });
-    });
-
-    it("has DECONSTRUCT method", () => {
-      const us = new UnknownValue("Test@1", "state");
-      expect(us[DECONSTRUCT]()).toEqual({
-        type: "Test@1",
-        state: "state",
-      });
-    });
-  });
-
-  // --------------------------------------------------------------------------
-  // ProblematicValue
-  // --------------------------------------------------------------------------
-
-  describe("ProblematicValue", () => {
-    it("preserves typeTag, state, and error", () => {
-      const ps = new ProblematicValue("BadType@1", { x: 1 }, "boom");
-      expect(ps.typeTag).toBe("BadType@1");
-      expect(ps.state).toEqual({ x: 1 });
-      expect(ps.error).toBe("boom");
-    });
-
-    it("has DECONSTRUCT method", () => {
-      const ps = new ProblematicValue("T@1", "s", "e");
-      expect(ps[DECONSTRUCT]()).toEqual({
-        type: "T@1",
-        state: "s",
-        error: "e",
-      });
-    });
-  });
-
-  // --------------------------------------------------------------------------
-  // ExplicitTagValue base class
-  // --------------------------------------------------------------------------
-
-  describe("ExplicitTagValue", () => {
-    it("UnknownValue is an instance of ExplicitTagValue", () => {
-      const us = new UnknownValue("Test@1", "state");
-      expect(us instanceof ExplicitTagValue).toBe(true);
-    });
-
-    it("ProblematicValue is an instance of ExplicitTagValue", () => {
-      const ps = new ProblematicValue("Test@1", "state", "oops");
-      expect(ps instanceof ExplicitTagValue).toBe(true);
-    });
-
-    it("ExplicitTagValue provides access to typeTag and state", () => {
-      const us: ExplicitTagValue = new UnknownValue("Tag@2", 42);
-      expect(us.typeTag).toBe("Tag@2");
-      expect(us.state).toBe(42);
-
-      const ps: ExplicitTagValue = new ProblematicValue(
-        "Bad@1",
-        "data",
-        "err",
-      );
-      expect(ps.typeTag).toBe("Bad@1");
-      expect(ps.state).toBe("data");
-    });
-  });
-
-  // --------------------------------------------------------------------------
   // JsonEncodingContext public API
   // --------------------------------------------------------------------------
 
@@ -1777,155 +1640,6 @@ describe("JsonEncodingContext", () => {
     it("lenient can be set to true", () => {
       const ctx = new JsonEncodingContext({ lenient: true });
       expect(ctx.lenient).toBe(true);
-    });
-  });
-
-  // --------------------------------------------------------------------------
-  // nativeFromFabricValueModern
-  // --------------------------------------------------------------------------
-
-  describe("nativeFromFabricValueModern", () => {
-    it("deeply unwraps FabricError in objects (frozen)", () => {
-      const err = new Error("deep");
-      const se = new FabricError(err);
-      const obj = {
-        error: se,
-        code: 500,
-      } as unknown as FabricValue;
-      const result = nativeFromFabricValueModern(obj) as Record<
-        string,
-        unknown
-      >;
-      expect(result.error).toBeInstanceOf(Error);
-      expect((result.error as Error).message).toBe("deep");
-      expect(Object.isFrozen(result.error)).toBe(true);
-      expect(result.code).toBe(500);
-      expect(Object.isFrozen(result)).toBe(true);
-    });
-
-    it("deeply unwraps FabricError in arrays (frozen)", () => {
-      const err = new Error("array");
-      const se = new FabricError(err);
-      const arr = [1, se, 3] as unknown as FabricValue;
-      const result = nativeFromFabricValueModern(arr) as unknown[];
-      expect(result[0]).toBe(1);
-      expect(result[1]).toBeInstanceOf(Error);
-      expect((result[1] as Error).message).toBe("array");
-      expect(Object.isFrozen(result[1])).toBe(true);
-      expect(result[2]).toBe(3);
-      expect(Object.isFrozen(result)).toBe(true);
-    });
-
-    it("output is not frozen when frozen=false", () => {
-      const obj = Object.freeze({
-        a: 1,
-        b: "two",
-      }) as unknown as FabricValue;
-      const result = nativeFromFabricValueModern(obj, false) as Record<
-        string,
-        unknown
-      >;
-      // Output should be a new, unfrozen object.
-      expect(Object.isFrozen(result)).toBe(false);
-      result.a = 99; // should not throw
-      expect(result.a).toBe(99);
-    });
-
-    it("output is frozen when frozen=true (default)", () => {
-      const obj = { a: 1, b: "two" } as unknown as FabricValue;
-      const result = nativeFromFabricValueModern(obj) as Record<
-        string,
-        unknown
-      >;
-      expect(Object.isFrozen(result)).toBe(true);
-    });
-
-    it("preserves sparse holes", () => {
-      const arr = new Array(3) as FabricValue[];
-      arr[0] = 1;
-      arr[2] = 3;
-      Object.freeze(arr);
-      const result = nativeFromFabricValueModern(
-        arr as FabricValue,
-      ) as unknown[];
-      expect(result.length).toBe(3);
-      expect(result[0]).toBe(1);
-      expect(1 in result).toBe(false); // hole preserved
-      expect(result[2]).toBe(3);
-    });
-
-    it("passes through non-native FabricInstance", () => {
-      const us = new UnknownValue("Test@1", null);
-      const obj = { thing: us } as unknown as FabricValue;
-      const result = nativeFromFabricValueModern(obj) as Record<
-        string,
-        unknown
-      >;
-      expect(result.thing).toBe(us);
-    });
-
-    it("deeply unwraps FabricMap to FrozenMap", () => {
-      const map = new Map<FabricValue, FabricValue>([
-        ["x", 10],
-      ] as [FabricValue, FabricValue][]);
-      const sm = new FabricMap(map);
-      const obj = { data: sm } as unknown as FabricValue;
-      const result = nativeFromFabricValueModern(obj) as Record<
-        string,
-        unknown
-      >;
-      expect(result.data).toBeInstanceOf(FrozenMap);
-      expect((result.data as Map<string, number>).get("x")).toBe(10);
-    });
-
-    it("deeply unwraps FabricSet to FrozenSet", () => {
-      const set = new Set<FabricValue>([42] as FabricValue[]);
-      const ss = new FabricSet(set);
-      const arr = [ss] as unknown as FabricValue;
-      const result = nativeFromFabricValueModern(arr) as unknown[];
-      expect(result[0]).toBeInstanceOf(FrozenSet);
-      expect((result[0] as Set<number>).has(42)).toBe(true);
-    });
-
-    it("deeply unwraps Error internals (C2)", () => {
-      // Error with a FabricError cause and a custom FabricMap property.
-      const innerErr = new Error("inner");
-      const innerSe = new FabricError(innerErr);
-      const outerErr = new Error("outer");
-      outerErr.cause = innerSe;
-      (outerErr as unknown as Record<string, unknown>).data = new FabricMap(
-        new Map([["k", 1]] as [FabricValue, FabricValue][]),
-      );
-      const outerSe = new FabricError(outerErr);
-
-      const result = nativeFromFabricValueModern(
-        outerSe as FabricValue,
-      ) as Error;
-      expect(result).toBeInstanceOf(Error);
-      expect(result.message).toBe("outer");
-      // cause should be deeply unwrapped to a native Error, not FabricError.
-      expect(result.cause).toBeInstanceOf(Error);
-      expect((result.cause as Error).message).toBe("inner");
-      // custom property should be unwrapped to FrozenMap.
-      const data = (result as unknown as Record<string, unknown>).data;
-      expect(data).toBeInstanceOf(FrozenMap);
-    });
-
-    it("deeply unwraps Error internals unfrozen (C2)", () => {
-      const innerErr = new Error("inner");
-      const innerSe = new FabricError(innerErr);
-      const outerErr = new Error("outer");
-      outerErr.cause = innerSe;
-      const outerSe = new FabricError(outerErr);
-
-      const result = nativeFromFabricValueModern(
-        outerSe as FabricValue,
-        false,
-      ) as Error;
-      expect(result).toBeInstanceOf(Error);
-      expect(Object.isFrozen(result)).toBe(false);
-      expect(result.cause).toBeInstanceOf(Error);
-      expect(Object.isFrozen(result.cause)).toBe(false);
     });
   });
